split lib modules into separate files
